requirements.txt
jax==0.1.72
jaxlib==0.1.51
scipy==1.4.1
numpy==1.18.5
matplotlib
pandas
seaborn
simple-parsing
cython
ipopt
gin-config
absl-py
gym
stable-baselines3
